Create author common functions
